+++ CHANGELOG.md
@@ -56,6 +56,14 @@ the release.
   ([#1857](https://github.com/open-telemetry/opentelemetry-demo/pull/1857))
 * [currency] rename currencyservice to currency
   ([#1858](https://github.com/open-telemetry/opentelemetry-demo/pull/1858))
+* [email] rename emailservice to email
+  ([#1861](https://github.com/open-telemetry/opentelemetry-demo/pull/1861))
+* [fraud-detection] rename frauddetectionservice to fraud-detection
+  ([#1862](https://github.com/open-telemetry/opentelemetry-demo/pull/1862))
+* [payment] rename paymentservice to payment
+  ([#1863](https://github.com/open-telemetry/opentelemetry-demo/pull/1863))
+* [recommendation] rename recommendationservice to recommendation
+  ([#1865](https://github.com/open-telemetry/opentelemetry-demo/pull/1865))
 
 ## 1.12.0